Generation config separation
