+++ tests/python_tests/test_continuous_batching.py
@@ -11,9 +11,7 @@ from pathlib import Path
 from openvino_genai import ContinuousBatchingPipeline, LLMPipeline, GenerationConfig, SchedulerConfig,  Tokenizer, draft_model
 
 from common import get_default_properties, get_hugging_face_models, convert_models, generate_and_compare_with_reference_text, \
-    get_scheduler_config, get_greedy, run_cb_pipeline_with_ref, get_beam_search, get_greedy, \
-    get_multinomial_all_parameters, get_multinomial_temperature_and_num_return_sequence, \
-    get_multinomial_temperature_and_top_k, get_multinomial_temperature, get_multinomial_temperature_and_top_p
+    get_scheduler_config, run_cb_pipeline_with_ref
 from test_sampling import RandomSamplingTestStruct, get_current_platform_ref_texts
 
 from ov_genai_test_utils import (
@@ -34,6 +32,10 @@ def read_models_list(file_name: str):
     return models
 
 from shutil import rmtree
+
+from utils.generation_config import get_greedy, get_beam_search, \
+    get_multinomial_all_parameters, get_multinomial_temperature_and_num_return_sequence, \
+    get_multinomial_temperature_and_top_k, get_multinomial_temperature, get_multinomial_temperature_and_top_p
 
 #
 # e2e tests on random and real models
